requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.122.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
oauthlib==3.3.1
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
